chore: make thank-you-page work without intermediate file
