+++ packages/gatsby-theme-project-portal/src/layouts/ThankYouPageLayout.tsx
@@ -1,7 +1,9 @@
 import React, { FunctionComponent } from "react"
-import { Link } from "gatsby"
+import { graphql, Link } from "gatsby"
 import { HeaderWithImage } from "../components"
 import { ImageDataLike } from "gatsby-plugin-image"
+
+export { Head } from "../hooks"
 
 interface ThankYouProps {
   data: {
@@ -40,3 +42,21 @@ export const ThankYouPageLayout: FunctionComponent<ThankYouProps> = ({
     </>
   )
 }
+
+export default ThankYouPageLayout
+
+export const query = graphql`
+  query ThankYouQuery($slug: String!) {
+    ...HeadData
+    ...LayoutData
+    page: generalPage(slug: { eq: $slug }) {
+      title
+      description: lede
+    }
+    generalPage(slug: { eq: $slug }) {
+      image {
+        ...HeaderWithImageBackground
+      }
+    }
+  }
+`
