--- packages/gatsby-theme-project-portal/src/templates/thank-you-page.js
@@ -1,21 +1,0 @@
-import { graphql } from "gatsby"
-import { ThankYouPageLayout } from "../layouts"
-
-export default ThankYouPageLayout
-export { Head } from "../hooks"
-
-export const query = graphql`
-  query ThankYouQuery($slug: String!) {
-    ...HeadData
-    ...LayoutData
-    page: generalPage(slug: { eq: $slug }) {
-      title
-      description: lede
-    }
-    generalPage(slug: { eq: $slug }) {
-      image {
-        ...HeaderWithImageBackground
-      }
-    }
-  }
-`
